% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{addToWorktable}
\alias{addToWorktable}
\title{add to worktable}
\usage{
addToWorktable(
  RackLabel = "",
  grid = 0,
  site = 0,
  RackType = "",
  Volume = 0,
  LiquidClass = ""
)
}
\arguments{
\item{RackLabel}{Max. 32 characters User-defined label (name) which is
assigned to the labware.}

\item{grid}{integer 1 - 67 labware location - carrier grid position}

\item{site}{integer 0 - 127 labware location - (site on carrier - 1)}

\item{RackType}{Max. 32 characters, Labware type (configuration name),
e.g. “384 Well, landscape”. Only used for basic worklists.}

\item{Volume}{needed in Rack: this will be filled up by the functions}

\item{LiquidClass}{class in the Rack.}
}
\value{
a string with warnings generated by the function
}
\description{
adds the labware to a worktable file to keep track of used labware
}
\seealso{
Other general methods: 
\code{\link{addToWorklist}()},
\code{\link{init}()},
\code{\link{write.gwl}()}
}
\concept{general methods}
