% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advGwlFunctions.R
\name{adv_gwl_comment}
\alias{adv_gwl_comment}
\title{Adv Comment Command}
\usage{
adv_gwl_comment(comment)
}
\description{
insert a comment into the advanced worklist shows up in evoware running log
(as opposed to \code{gwl_comment} which only shows up in the .gwl file).
}
\seealso{
Other advanced worklist commands: \code{\link{MCArelative}},
  \code{\link{MCAvector}}, \code{\link{adv_aspirate}},
  \code{\link{adv_dispense}}, \code{\link{adv_mix}},
  \code{\link{adv_wash}}, \code{\link{fwCommand}},
  \code{\link{moveLiHa}}, \code{\link{moveMCA}},
  \code{\link{startTimer}}, \code{\link{waitTimer}}
}
\concept{advanced worklist commands}
