% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advGwlFunctions.R
\name{adv_mix}
\alias{adv_mix}
\title{Adv Mix Command}
\usage{
adv_mix(tipMask = c(1, 0, 0, 0, 0, 0, 0, 0),
  liquidClass = gwl$defaults$LiquidClass, volumes = gwl$defaults$Volume,
  RackLabel, spacing = 2, wellSelection, ncol, nrow, cycles,
  noOfLoopOptions = 0, loopName = "", action = 0, difference = 0,
  arm = 0)
}
\arguments{
\item{tipMask}{\code{vector("integer",8L)}  containing 0, 1 indicating
selected tips (e.g. \code{c(1,0,1,0,0,0,0,0)} = pins 1 + 3)}

\item{liquidClass}{string1 liquid class name}

\item{volumes}{\code{vector("integer",12L)} 12 expressions which specify the
volume for each tip. Specify 0 for tips which are not used or
not fitted.}

\item{spacing}{integer 1 -  .. tip spacing, must be 2 for 384 well plates}

\item{wellSelection}{vector containing the wells from which the LiHa should
aspirate (should be in one column!)}

\item{ncol}{integer, number of colums (left to right on worktable) of labware}

\item{nrow}{integer, number of rows (back to front on worktable) of labware}

\item{cycles}{integer 0 .. 100 number of mixing cycles}

\item{noOfLoopOptions}{integer 0 - 7 number of loop options required. If
non-zero, parameters \code{loopname, action, difference} are included
n times, where n = \code{noOfLoopOptions}}

\item{loopName}{loop name, \code{vector("character",noOfLoopOptions)}}

\item{action}{integer 0 - 3 loop action: 0 = vary the column,
1=vary the row 2=vary the well 3=vary the rack,
 \code{vector("character",noOfLoopOptions)}}

\item{difference}{integer number (“vary by”) which is used for the loop
action, \code{vector("character",noOfLoopOptions)}}

\item{arm}{integer 0 - 1 number of the LiHa performing the action:
0 = LiHa 1=LiHa 2, should always be 0}

\item{p_RackLabel}{Max. 32 characters User-defined label (name) which is
assigned to the labware.}

\item{grid}{integer 1 - 67 labware location - carrier grid position}

\item{site}{integer 0 - 127 labware location - (site on carrier - 1)}
}
\description{
Adv Mix Command
}
\seealso{
Other advanced worklist commands: \code{\link{MCArelative}},
  \code{\link{MCAvector}}, \code{\link{adv_aspirate}},
  \code{\link{adv_dispense}},
  \code{\link{adv_gwl_comment}}, \code{\link{adv_wash}},
  \code{\link{fwCommand}}, \code{\link{moveLiHa}},
  \code{\link{moveMCA}}, \code{\link{startTimer}},
  \code{\link{waitTimer}}
}
\concept{advanced worklist commands}
