% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advGwlFunctions.R
\name{adv_wash}
\alias{adv_wash}
\title{Adv Wash Command}
\usage{
adv_wash(tipMask = rep(1, 8), RackLabel_waste = "wash1_waste",
  RackLabel_cleaner = "wash1_shallow", wasteVol = 3,
  wasteDelay = 500, cleanerVol = 4, cleanerDelay = 500,
  Airgap = 10, airgapSpeed = 70, retractSpeed = 30, FastWash = 1,
  lowVolume = 0, arm = 0)
}
\arguments{
\item{tipMask}{\code{vector("integer",8L)}  containing 0, 1 indicating
selected tips (e.g. \code{c(1,0,1,0,0,0,0,0)} = pins 1 + 3)}

\item{RackLabel_waste}{Name of the waste rack}

\item{RackLabel_cleaner}{Name of the cleaner rack}

\item{wasteVol}{integer 0 - 100 volume in waste in ml}

\item{wasteDelay}{integer 0 - 1000 delay before closing valves in waste in ms}

\item{cleanerVol}{expression 0 - 100 volume in cleaner in ml}

\item{cleanerDelay}{integer 0 - 1000 delay before closing valves in cleaner
in ms}

\item{Airgap}{floating point 0 - 100 volume of airgap in µl which is
aspirated after washing the tips (system trailing airgap)}

\item{airgapSpeed}{integer 1 - 1000 speed of airgap aspiration in µl/s}

\item{retractSpeed}{integer 1 - 100 retract speed in mm/s}

\item{FastWash}{integer 0 - 1 1 = use fast-wash-module}

\item{lowVolume}{integer 0 - 1 1 = use pinch valves}

\item{arm}{integer 0 - 1 number of the LiHa performing the action:
0 = LiHa 1=LiHa 2, should always be 0}
}
\description{
Adv Wash Command
}
\seealso{
Other advanced worklist commands: \code{\link{MCArelative}},
  \code{\link{MCAvector}}, \code{\link{adv_aspirate}},
  \code{\link{adv_dispense}},
  \code{\link{adv_gwl_comment}}, \code{\link{adv_mix}},
  \code{\link{fwCommand}}, \code{\link{moveLiHa}},
  \code{\link{moveMCA}}, \code{\link{startTimer}},
  \code{\link{waitTimer}}
}
\concept{advanced worklist commands}
