% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_OutDet.R
\name{all_OutDet}
\alias{all_OutDet}
\title{Apply All Outlier Detection Threshold Methods on One Dataset}
\usage{
all_OutDet(name, x, w, Q, alpha = 0.01)
}
\arguments{
\item{name}{String. A label for the dataset (e.g., "abide1")}

\item{x}{High-kurtosis matrix (T × K)}

\item{w}{Low-kurtosis matrix (T × L)}

\item{Q}{Integer. Number of variables in x (usually ncol(x))}

\item{alpha}{Numeric. Quantile for threshold (default 0.01)}
}
\value{
A list containing thresholds for each method
}
\description{
Apply All Outlier Detection Threshold Methods on One Dataset
}
