% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_args.R
\name{alpha}
\alias{alpha}
\title{Alpha parameter documentation}
\arguments{
\item{alpha}{Significance level used to compute RD threshold (default = 0.01 for 99th percentile).}
}
\description{
Alpha parameter documentation
}
