% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accFunctions.R
\name{ANSIIfromVector}
\alias{ANSIIfromVector}
\title{ANSII from Vector}
\usage{
ANSIIfromVector(x, length = 7)
}
\arguments{
\item{x}{vector containing a selection of the numbers 1-7}

\item{length}{length of binary number, should always be 7.}
}
\description{
generates ANSII letter from a vector with numbers 1 to 7 indicating which well in this group of 7 is used.
}
\seealso{
Other accessory functions: \code{\link{ANSIIFromWells}},
  \code{\link{wellsFromANSII}}
}
