% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accFunctions.R
\name{ANSIIFromWells}
\alias{ANSIIFromWells}
\title{ANSII String from Wells}
\usage{
ANSIIFromWells(wellSelection, ncol = 24, nrow = 16)
}
\arguments{
\item{wellSelection}{Well position(s) in the labware. The position starts with 1 and
increases from rear to front and left to right.}

\item{ncol}{Number of columns in labware, defaults to 24 (a 384 well plate)}

\item{nrow}{Number of rows in labware, defaults to 16 (a 384 well plate)}
}
\description{
calculates string for well selection in advanced worklist commands
}
\seealso{
Other accessory functions: \code{\link{ANSIIfromVector}},
  \code{\link{wellsFromANSII}}
}
