% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accFunctions.R
\name{binToInt}
\alias{binToInt}
\title{binary string to integer}
\usage{
binToInt(x)
}
\arguments{
\item{x}{character representation of a binary number, eg. "1101" (=13)}
}
\description{
no description
}
\section{Output}{

   an integer
}

\seealso{
Other hospital_experiment: \code{\link{intToBin}}
}
