% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_args.R
\name{boot_quant}
\alias{boot_quant}
\title{Boot_quant parameter documentation}
\arguments{
\item{boot_quant}{Numeric; confidence level for bootstrap confidence intervals (default = 0.95, for 95\% CI).}
}
\description{
Boot_quant parameter documentation
}
