% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_args.R
\name{cov_mcd}
\alias{cov_mcd}
\title{Cov_mcd parameter documentation}
\arguments{
\item{cov_mcd}{Optional covariance matrix (p × p); required in "manual" mode.}
}
\description{
Cov_mcd parameter documentation
}
