% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_args.R
\name{cutoff}
\alias{cutoff}
\title{Cutoff parameter documentation}
\arguments{
\item{cutoff}{A numeric value indicating how many MADs away from the median to flag as outliers. The default value is set to be 4.}
}
\description{
Cutoff parameter documentation
}
