% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwlFunctions.R
\name{dispense}
\alias{dispense}
\title{Dispense Command}
\usage{
dispense(RackLabel = gwl$defaults$RackLabel,
  RackID = gwl$defaults$RackID, RackType = gwl$defaults$RackType,
  Position = gwl$defaults$Position, TubeID = gwl$defaults$TubeID,
  Volume = gwl$defaults$Volume, LiquidClass = gwl$defaults$LiquidClass,
  TipType = "", TipMask = gwl$defaults$TipMask,
  ForcedRackType = gwl$defaults$ForcedRackType,
  MinDetectedVolume = gwl$defaults$MinDetectedVolume)
}
\arguments{
\item{RackLabel}{Max. 32 characters User-defined label (name) which is assigned to the labware.}

\item{RackID}{Max. 32 characters, Labware barcode.}

\item{RackType}{Max. 32 characters, Labware type (configuration name), e.g. “384 Well, landscape”.}

\item{Position}{1 .. number of wells, Well position in the labware. The position starts with 1 and increases from rear to front and left to right.}

\item{TubeID}{Max. 32 characters, Tube barcode.}

\item{Volume}{0 .. +7158278, Pipetting volume in µl.}

\item{LiquidClass}{Max. 32 characters, This optional parameter overwrites the liquid class specified in the Worklist command.}

\item{TipType}{Reserved, must be omitted.}

\item{TipMask}{1 .. 128, This optional parameter specifies the tip you want to use. The tip number is bit-coded, i.e. Tip 1 = 1, Tip 2 = 2, Tip 3 = 4; Tip 4 = 8 and so on. Please note that you can only use a tip which has been enabled with Tip Selection in the Worklist command. If you do not specify TipMask, Freedom EVOware uses one of the tips specified with Tip Selection in the Worklist command.}

\item{ForcedRackType}{Max. 32 characters, This optional parameter is the configuration name of the labware. If specified, it is used instead of the labware type which is found on the Freedom EVOware worktable.}

\item{MinDetectedVolume}{0 .. +7158278 Liquid volume in µl. If this optional parameter is specified and Liquid Level Detection is enabled in the selected liquid class, it is used by the LiHa to determine the minimum liquid height which must be available in the well. Otherwise a liquid detection error (not enough liquid) will be triggered. MinDetectedVolume specifies the usable volume of liquid. If the aspiration position in the liquid class specifies tip immersion (tip submerging), the actual liquid height in the well must be higher to allow immersion (i.e. the volume of liquid in the well must be more than MinDetectedVolume).}
}
\description{
Dispense Command
}
\seealso{
Other basic worklist commands: \code{\link{aspirate}},
  \code{\link{distribute}}, \code{\link{gwl_break}},
  \code{\link{gwl_comment}}, \code{\link{gwl_flush}},
  \code{\link{wash}}
}
\concept{basic worklist commands}
