% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_args.R
\name{dist}
\alias{dist}
\title{Dist parameter documentation}
\arguments{
\item{dist}{Logical; if TRUE, compute squared robust Mahalanobis distances for all observations.}
}
\description{
Dist parameter documentation
}
