% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwlFunctions.R
\name{distribute}
\alias{distribute}
\title{Reagent Distribution Command}
\usage{
distribute(SrcRackLabel = "", SrcRackID = "", SrcRackType = "",
  SrcPosStart = "", SrcPosEnd = "", DestRackLabel = "",
  DestRackID = "", DestRackType = "", DestPosStart = "",
  DestPosEnd = "", Volume = gwl$defaults$Volume,
  LiquidClass = gwl$defaults$LiquidClass, NoOfDitiReuses = 1,
  NoOfMultiDisp = 1, Direction = 0, ExcludeDestWell = NA)
}
\arguments{
\item{SrcRackLabel}{Max. 32 characters User-defined label (name) which is assigned to the source labware.}

\item{SrcRackID}{Max. 32 characters Source labware barcode.}

\item{SrcRackType}{Max. 32 characters Source labware type (configuration name), e.g. “384 Well, landscape”.}

\item{SrcPosStart}{1 .. number of wells First well to be used in the source labware.}

\item{SrcPosEnd}{1 .. number of wells Last well to be used in the source labware.}

\item{DestRackLabel}{Max. 32 characters User-defined label (name) which is assigned to the destination labware.}

\item{DestRackID}{Max. 32 characters Destination labware barcode.}

\item{DestRackType}{Max. 32 characters Destination labware type (configuration name), e.g. “384 Well, landscape”.}

\item{DestPosStart}{1 .. number of wells First well to be used in the destination labware.}

\item{DestPosEnd}{1 .. number of wells Last well to be used in the destination labware.}

\item{Volume}{0 .. +7158278 Dispense volume in the destination labware in µl.}

\item{LiquidClass}{Max. 32 characters This optional parameter overwrites the liquid class specified in the Worklist command. #' @param NoOfDitiReuses Optional maximum number of DiTi reuses allowed (default 1 = no DiTi reuse).}

\item{NoOfMultiDisp}{Optional maximum number of dispenses in a multi-dispense sequence (default 1 = no multi-dispense).}

\item{Direction}{Optional pipetting direction (0 = left to right, 1 = right to left; default = 0).}

\item{ExcludeDestWell}{Optional vector of wells in destination labware to be excluded from pipetting (e.g. c(23,44,36))}
}
\description{
The Reagent Distribution record specifies the aspiration and dispensing parameters for all of the tips you have chosen (and dispenses the same volume in all of the chosen wells in the destination labware).
}
\seealso{
Other basic worklist commands: \code{\link{aspirate}},
  \code{\link{dispense}}, \code{\link{gwl_break}},
  \code{\link{gwl_comment}}, \code{\link{gwl_flush}},
  \code{\link{wash}}
}
\concept{basic worklist commands}
