% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accFunctions.R
\name{.advVolumeString}
\alias{.advVolumeString}
\title{Volume string generation}
\usage{
.advVolumeString(volumes, tipMask)
}
\arguments{
\item{volumes}{\code{vector("numeric",12L)}}
}
\description{
generates the string of volumes needed for the advanced worklist
}
\seealso{
Other internal: \code{\link{.binToDec}},
  \code{\link{.convertXMLtime}}, \code{\link{.decToBin}},
  \code{\link{.listToDf}}, \code{\link{.updateVolume}},
  \code{\link{getGridSite}}
}
\concept{internal}
