% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accFunctions.R
\name{.binToDec}
\alias{.binToDec}
\title{binary to dec}
\usage{
.binToDec(x)
}
\arguments{
\item{x}{a vector consisting of 0 and 1 representing a binary number}
}
\description{
takes a vector representing a binary number and gives the decimal value
}
\seealso{
Other internal: 
\code{\link{.advVolumeString}()},
\code{\link{.convertXMLtime}()},
\code{\link{.decToBin}()},
\code{\link{.listToDf}()},
\code{\link{.updateVolume}()},
\code{\link{getGridSite}()}
}
\concept{internal}
