% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accFunctions.R
\name{.convertXMLtime}
\alias{.convertXMLtime}
\title{Convert time from Reader XML}
\usage{
.convertXMLtime(string)
}
\arguments{
\item{string}{string of time in format provided by tecan infinity M200
or list_lengths thereof}
}
\description{
converts the time from the reader xml to a R format
}
\section{Output}{

  \code{POSIX.ct} timestamp or a vector thereof
}

\seealso{
Other internal: \code{\link{.advVolumeString}},
  \code{\link{.binToDec}}, \code{\link{.decToBin}},
  \code{\link{.listToDf}}, \code{\link{.updateVolume}},
  \code{\link{getGridSite}}
}
\concept{internal}
