% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accFunctions.R
\name{.decToBin}
\alias{.decToBin}
\title{dec to binary}
\usage{
.decToBin(x, l = 7)
}
\arguments{
\item{x}{a decimal number}

\item{l}{length of the vector}
}
\description{
takes a decimal number and returns a vector representing a binary number.
used for adv. well selection, hence length=7
}
\seealso{
Other internal: \code{\link{.advVolumeString}},
  \code{\link{.binToDec}}, \code{\link{.convertXMLtime}},
  \code{\link{.listToDf}}, \code{\link{.updateVolume}},
  \code{\link{getGridSite}}
}
\concept{internal}
