% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accFunctions.R
\name{.listToDf}
\alias{.listToDf}
\title{List to data.frame}
\usage{
.listToDf(list)
}
\arguments{
\item{list}{a list of vectors}
}
\description{
takes a list of vectors (possibly named) and makes a data.frame for easier
display. Trys to make guesses for column names. Used in \code{\link{read_Infinite_XML}}
}
\seealso{
Other internal: 
\code{\link{.advVolumeString}()},
\code{\link{.binToDec}()},
\code{\link{.convertXMLtime}()},
\code{\link{.decToBin}()},
\code{\link{.updateVolume}()},
\code{\link{getGridSite}()}
}
\concept{internal}
