% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accFunctions.R
\name{.updateVolume}
\alias{.updateVolume}
\title{add/subtract Volumes on worktable}
\usage{
.updateVolume(volumes, RackLabel, direction = 1)
}
\arguments{
\item{volumes}{Volumes}

\item{RackLabel}{name of the Racklabel}

\item{direction}{1 to add volume, -1 to subtract volume}
}
\description{
internal function to add/subtract volumes from labware
}
\seealso{
Other internal: 
\code{\link{.advVolumeString}()},
\code{\link{.binToDec}()},
\code{\link{.convertXMLtime}()},
\code{\link{.decToBin}()},
\code{\link{.listToDf}()},
\code{\link{getGridSite}()}
}
\concept{internal}
