% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procedures.R
\name{evowait}
\alias{evowait}
\title{wait for t seconds}
\usage{
evowait(timer, timeSpan)
}
\arguments{
\item{timer}{Number of Timer to (re-)start}

\item{timeSpan}{duration [0.02 - 86400]}
}
\description{
simpler timer
}
\seealso{
Other advanced worklist procedures: \code{\link{iPTdeactivate}},
  \code{\link{iPTdip}}, \code{\link{sterile_wash_TT}},
  \code{\link{sterile_wash_phage}},
  \code{\link{sterile_wash}}, \code{\link{tiptouch}},
  \code{\link{washPinTool}}
}
