% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advGwlFunctions.R
\name{fwCommand}
\alias{fwCommand}
\title{Send Firmware command}
\usage{
fwCommand(
  command,
  options = 1,
  parameter = 1,
  response = "",
  error = "",
  errorScope = 2,
  responseScope = 2,
  responseType = 0
)
}
\arguments{
\item{command}{string, the command}

\item{options}{int 0-7, bit-coded options: 1 = wait for termination of
command, 2 = store error code in variable, 4 = store answer in variable}

\item{parameter}{int 1-16 number of the parameter in the response which is
stored in the variable “response”}

\item{response}{string, variable to store response, empty for no storage}

\item{error}{string, variable to store error, empty for no storage}

\item{errorScope}{int 0-2, scope of error variable: 0 = Run, 1 = Instance,
2 = Script}

\item{responseScope}{int 0-2, scope of response variable: 0 = Run,
1 = Instance, 2 = Script}

\item{responseType}{int 0-1, type of response variable. 0 = numeric,
1 = string}
}
\description{
sends a firmware command
}
\seealso{
Other advanced worklist commands: 
\code{\link{MCArelative}()},
\code{\link{MCAvector}()},
\code{\link{adv_aspirate}()},
\code{\link{adv_dispense}()},
\code{\link{adv_gwl_comment}()},
\code{\link{adv_mix}()},
\code{\link{adv_wash}()},
\code{\link{moveLiHa}()},
\code{\link{moveMCA}()},
\code{\link{startTimer}()},
\code{\link{waitTimer}()}
}
\concept{advanced worklist commands}
