% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accFunctions.R
\name{generateMasks}
\alias{generateMasks}
\title{generateMasks}
\usage{
generateMasks(tip, vol)
}
\arguments{
\item{tip}{a vector of tips}

\item{vol}{a vector of volumes for each of the tips}
}
\description{
tip & volumen masks from wells and volumes
}
\seealso{
Other general: 
\code{\link{rc_to_well}()},
\code{\link{read_storex}()},
\code{\link{shortestPath}()}
}
\concept{general}
