% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accFunctions.R
\name{getGridSite}
\alias{getGridSite}
\title{get grid/site from racklabel}
\usage{
getGridSite(RackLabel)
}
\arguments{
\item{RackLabel}{name of the Rack}
}
\description{
internal function to get grid/site from racklabel.
Used in advanced worklist functions
}
\seealso{
Other internal: 
\code{\link{.advVolumeString}()},
\code{\link{.binToDec}()},
\code{\link{.convertXMLtime}()},
\code{\link{.decToBin}()},
\code{\link{.listToDf}()},
\code{\link{.updateVolume}()}
}
\concept{internal}
