% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwlFunctions.R
\name{gwl_comment}
\alias{gwl_comment}
\title{Comment Command}
\usage{
gwl_comment(comment)
}
\description{
insert a comment into the worklist (ignored by EVOware).
}
\seealso{
Other basic worklist commands: 
\code{\link{aspirate}()},
\code{\link{dispense}()},
\code{\link{distribute}()},
\code{\link{gwl_break}()},
\code{\link{gwl_flush}()},
\code{\link{wash}()}
}
\concept{basic worklist commands}
