% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwlCheckFunctions.R
\name{gwl_df_part}
\alias{gwl_df_part}
\title{gwl data_frame part}
\usage{
gwl_df_part(type = as.character(NA), command = as.character(NA),
  attributes = as.character(NA), tipMask = as.character(NA),
  LC = as.character(NA), volumes = as.character(NA), grid = NA,
  site = NA, spacing = NA, ncol = NA, nrow = NA,
  wellSelection = as.character(NA), zMove = NA, zTarget = NA,
  offset = NA, speed = NA, loop = as.character(NA),
  arm = as.character(NA))
}
\description{
helper function to create a dataframe of a gwl line
}
\seealso{
Other gwl_check: \code{\link{read_gwl_df}},
  \code{\link{read_gwl}}
}
\concept{gwl_check}
