% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwlFunctions.R
\name{gwl_flush}
\alias{gwl_flush}
\title{Flush Command}
\usage{
gwl_flush()
}
\description{
It discards the contents of the tips without washing them.
}
\seealso{
Other basic worklist commands: \code{\link{aspirate}},
  \code{\link{dispense}}, \code{\link{distribute}},
  \code{\link{gwl_break}}, \code{\link{gwl_comment}},
  \code{\link{wash}}
}
