% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwlCheckFunctions.R
\name{gwlToDataFrame}
\alias{gwlToDataFrame}
\title{crate data.frames from worklist}
\usage{
gwlToDataFrame(worklist, path = ".", .progress = TRUE)
}
\arguments{
\item{worklist}{worklist, read in with read_gwl}

\item{path}{where should the tables be saved}

\item{.progress}{should a progress bar be shown?}
}
\description{
generates data.frame combining commands if they are the same
}
\section{Output}{

 a lsit of pretty knitr::kable tables with images
 images are safed in a subfolder worklist_img_temp
}

\seealso{
Other gwl check: \code{\link{gwlToHTML}},
  \code{\link{gwlToTable}}
}
\concept{gwl check}
