% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwlCheckFunctions.R
\name{gwlToHTML}
\alias{gwlToHTML}
\title{crate a html report from a worklist file}
\usage{
gwlToHTML(gwl = file.choose(), output = NULL, .progress = TRUE, cleanup = TRUE)
}
\arguments{
\item{output}{where should the output be saved? If \code{NULL} then the html
is saved back to where the worklist lives.}

\item{worklist}{path to worklist, defaults to \code{file.choose()}}
}
\description{
generate pretty html report
}
\section{Output}{

 generates html with pretty knitr::kable tables with images
}

\seealso{
Other gwl check: 
\code{\link{gwlToDataFrame}()},
\code{\link{gwlToTable}()}
}
\concept{gwl check}
