% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwlCheckFunctions.R
\name{gwlToTable}
\alias{gwlToTable}
\title{crate a knitr::kable tables with images from worklist}
\usage{
gwlToTable(worklist, path = ".", .progress = TRUE)
}
\arguments{
\item{worklist}{worklist, read in with read_gwl}

\item{path}{where should the tables be saved}

\item{.progress}{should a progress bar be shown?}
}
\description{
generate pretty tables
}
\section{Output}{

 a lsit of pretty knitr::kable tables with images
 images are safed in a subfolder worklist_img_temp
}

\seealso{
Other gwl check: \code{\link{gwlToDataFrame}},
  \code{\link{gwlToHTML}}
}
