% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_args.R
\name{imp_datasets}
\alias{imp_datasets}
\title{Imp_datasets parameter documentation}
\arguments{
\item{imp_datasets}{A list of M numeric matrices (T × p); multiply imputed datasets.}
}
\description{
Imp_datasets parameter documentation
}
