% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_args.R
\name{impute_method}
\alias{impute_method}
\title{Impute_method parameter documentation}
\arguments{
\item{impute_method}{Character string; imputation method for univariate outliers.}
}
\description{
Impute_method parameter documentation
}
