% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_args.R
\name{ind_incld}
\alias{ind_incld}
\title{Ind_incld parameter documentation}
\arguments{
\item{ind_incld}{Optional vector of row indices used to compute the robust mean; required in "manual" mode.}
}
\description{
Ind_incld parameter documentation
}
