% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{init}
\alias{init}
\title{Initialize Worklist & set defaults}
\usage{
init(filename = paste0(format(Sys.time(), "\%y\%m\%d"), "_gwl"),
  RackLabel = "trough200_1_frt", RackID = "", RackType = "",
  Position = 0, TubeID = "", Volume = 0, LiquidClass = "",
  TipMask = 0, ForcedRackType = "", MinDetectedVolume = "", Wash = "W",
  WTtemplate = "basic")
}
\arguments{
\item{filename}{string, the name of the generated worklist and worktable
(+ "_worktable.txt")}

\item{RackLabel}{Max. 32 characters User-defined label (name) which is
assigned to the labware.}

\item{RackID}{Max. 32 characters, Labware barcode.}

\item{RackType}{Max. 32 characters, Labware type (configuration name),
e.g. “384 Well, landscape”.}

\item{Position}{1 .. number of wells, Well position in the labware. The
position starts with 1 and increases from rear to front and left to right.}

\item{TubeID}{Max. 32 characters, Tube barcode.}

\item{Volume}{0 .. +7158278, Pipetting volume in µl.}

\item{LiquidClass}{Max. 32 characters, This optional parameter overwrites the
liquid class specified in the Worklist command.}

\item{TipMask}{1 .. 128, This optional parameter specifies the tip you want
to use. The tip number is bit-coded, i.e. Tip 1 = 1, Tip 2 = 2,
Tip 3 = 4; Tip 4 = 8 and so on. Please note that you can only use a tip
which has been enabled with Tip Selection in the Worklist command.
If you do not specify TipMask, Freedom EVOware uses one of the tips
specified with Tip Selection in the Worklist command.}

\item{ForcedRackType}{Max. 32 characters, This optional parameter is the
configuration name of the labware. If specified, it is used instead
of the labware type which is found on the Freedom EVOware worktable.}

\item{MinDetectedVolume}{0 .. +7158278 Liquid volume in µl. If this optional
parameter is specified and Liquid Level Detection is enabled in the
selected liquid class, it is used by the LiHa to determine the minimum
liquid height which must be available in the well. Otherwise a liquid
detection error (not enough liquid) will be triggered. MinDetectedVolume
specifies the usable volume of liquid. If the aspiration position in the
liquid class specifies tip immersion (tip submerging), the actual liquid
height in the well must be higher to allow immersion (i.e. the volume of
liquid in the well must be more than MinDetectedVolume).}

\item{Wash}{Default wash step to use}

\item{WTtemplate}{character, data.frame or NULL. character: worktable
template to load ("basic" only atm), data.frame: a worktable layout as a
data.frame, NULL: an empty worktable.}
}
\description{
Initialization. Generates object \code{gwl} in \code{.GlobalEnv}
}
\section{Output}{

  invisible(gwl), also saves object \code{gwl} to \code{.GlobalEnv}
}

\seealso{
Other general methods: \code{\link{addToWorklist}},
  \code{\link{addToWorktable}}, \code{\link{write.gwl}}
}
\concept{general methods}
