% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accFunctions.R
\name{intToBin}
\alias{intToBin}
\title{integer to binary string}
\usage{
intToBin(x, l = 4)
}
\arguments{
\item{x}{integer}

\item{l}{length of output (will be padded with 0)}
}
\description{
no description
}
\section{Output}{

   a string
}

\seealso{
Other hospital_experiment: \code{\link{binToInt}}
}
\concept{hospital_experiment}
