% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procedures.R
\name{iPTdip}
\alias{iPTdip}
\title{iPT384 pin transfer}
\usage{
iPTdip(rack, well, n_dips = 3, speed_move = 35, speed_in = 5, speed_out = 5)
}
\arguments{
\item{rack}{string, Racklabel of target, as in gwl$worktable}

\item{well}{int 1:384, which well should be hit with pin 1}

\item{n_dips}{number of dips in plate}

\item{speed_move}{movement speed up to z-dispense}

\item{speed_in}{movement speed when dipping}

\item{speed_out}{movement speed when retracting}
}
\description{
pick up 4 adjacent wells on a 384 well plate with the 4 pins in the back left
 corner of the iPT384. Make sure all other pins are deactivated!
}
\seealso{
Other advanced worklist procedures: 
\code{\link{evowait}()},
\code{\link{iPTdeactivate}()},
\code{\link{sterile_wash_TT}()},
\code{\link{sterile_wash_c}()},
\code{\link{sterile_wash_phage}()},
\code{\link{sterile_wash}()},
\code{\link{tiptouch}()},
\code{\link{washPinTool}()}
}
\concept{advanced worklist procedures}
