% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advGwlFunctions.R
\name{MCArelative}
\alias{MCArelative}
\title{MCA Relative movement}
\usage{
MCArelative(col_offset = 0, row_offset = 0, arm = 0, verbose = FALSE)
}
\arguments{
\item{col_offset}{integer, [-23,23], positive numbers move the MCA to the
right, negative to the left. Defaults to 0 = reset offset.}

\item{row_offset}{integer, [-15,15], positive numbers move the MCA to the
front, negative to the back. Defaults to 0 = reset offset.}

\item{arm}{integer 0 - 1, number of the MCA96 performing the action:
0 = MCA 1=MCA 2, should always be 0}

\item{verbose}{logical, print comments?}
}
\description{
Sets an offset to the MCA
}
\seealso{
Other advanced worklist commands: \code{\link{MCAvector}},
  \code{\link{adv_aspirate}}, \code{\link{adv_dispense}},
  \code{\link{adv_gwl_comment}}, \code{\link{adv_mix}},
  \code{\link{adv_wash}}, \code{\link{fwCommand}},
  \code{\link{moveLiHa}}, \code{\link{moveMCA}},
  \code{\link{startTimer}}, \code{\link{waitTimer}}
}
