% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_args.R
\name{method_univOut}
\alias{method_univOut}
\title{Method_univOut parameter documentation}
\arguments{
\item{method}{Character string. One of \code{"SHASH"} or \code{"robZ"}.}
}
\description{
Method_univOut parameter documentation
}
