% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MImpute_old.R
\name{MImpute_old}
\alias{MImpute_old}
\title{Multiple Imputation for High-Kurtosis ICA Components}
\usage{
MImpute_old(x, w, outlier_matrix, M = 50, k = 100)
}
\arguments{
\item{x}{A numeric matrix (n_time × p) of high-kurtosis ICA components to be imputed.}

\item{w}{A numeric matrix (n_time × L) of low-kurtosis ICA components used as predictors (required for MI).}

\item{outlier_matrix}{A logical matrix (same dim as x) indicating univariate outliers to be imputed.}

\item{M}{Integer; number of multiply imputed datasets (default = 5).}

\item{k}{Integer; number of perturbation cycles per imputation (default = 10).}
}
\value{
A list with:
\describe{
\item{\code{imp_datasets}}{List of M imputed versions of x}
\item{\code{outlier_matrix}}{Logical matrix of imputed outlier positions}
}
}
\description{
Performs multiple imputation using perturbed robust regression models.
}
