% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advGwlFunctions.R
\name{moveLiHa}
\alias{moveLiHa}
\title{Move LiHa}
\usage{
moveLiHa(tipMask = c(1, 0, 0, 0, 0, 0, 0, 0), RackLabel, spacing = 1,
  wellSelection, ncol, nrow, zMove, zTarget, offset, speed,
  noOfLoopOptions = 0, loopName = "", action = 0, difference = 0,
  arm = 0)
}
\arguments{
\item{tipMask}{\code{vector("integer",8L)}  containing 0, 1 indicating
selected tips (e.g. \code{c(1,0,1,0,0,0,0,0)} = pins 1 + 3)}

\item{RackLabel}{Max. 32 characters User-defined label (name) which is
assigned to the labware.}

\item{spacing}{integer 1 -  .. tip spacing, must be 2 for 384 well plates}

\item{wellSelection}{vector containing the wells from which the LiHa should
aspirate (should be in one column!)}

\item{ncol}{integer, number of colums (left to right on worktable) of labware}

\item{nrow}{integer, number of rows (back to front on worktable) of labware}

\item{zMove}{integer 0 - 4 type of movement: 0 = positioning with global
z-travel 1 = positioning with local z-travel 2 = x-move 3 = y-move
4 = z-move}

\item{zTarget}{integer 0 - 4 z-position after move: 0 = z-travel
1 = z-dispense 2 = z-start 3 = z-max 4 = global z-travel}

\item{offset}{floating point [-1000,1000] offset in mm added to z-position
(zTarget)}

\item{speed}{floating point [0.1,400] move speed in mm/s if zMove is 2,3 or 4}

\item{noOfLoopOptions}{integer 0 - 7 number of loop options required. If
non-zero, parameters \code{loopname, action, difference} are included
n times, where n = \code{noOfLoopOptions}}

\item{loopName}{loop name, \code{vector("character",noOfLoopOptions)}}

\item{action}{integer 0 - 3 loop action: 0 = vary the column,
1=vary the row, 2=vary the well 3=vary the rack,
\code{vector("character",noOfLoopOptions)}}

\item{difference}{integer number (“vary by”) which is used for the loop
action, \code{vector("character",noOfLoopOptions)}}

\item{arm}{integer 0 - 1 number of the LiHa performing the action:
0 = LiHa 1=LiHa 2, should always be 0}
}
\description{
Move LiHa
}
\seealso{
Other advanced worklist commands: \code{\link{MCArelative}},
  \code{\link{MCAvector}}, \code{\link{adv_aspirate}},
  \code{\link{adv_dispense}},
  \code{\link{adv_gwl_comment}}, \code{\link{adv_mix}},
  \code{\link{adv_wash}}, \code{\link{fwCommand}},
  \code{\link{moveMCA}}, \code{\link{startTimer}},
  \code{\link{waitTimer}}
}
