% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advGwlFunctions.R
\name{moveMCA}
\alias{moveMCA}
\title{MCA Move}
\usage{
moveMCA(RackLabel, quadrant = 1, ncol = 24, nrow = 16, zMove, zTarget,
  offset = 0, speed, noOfLoopOptions = 0, loopName = "", action = 0,
  difference = 0, arm = 0)
}
\arguments{
\item{RackLabel}{Max. 32 characters User-defined label (name) which is
assigned to the labware.}

\item{quadrant}{integer 1-4, which quadrant?}

\item{ncol}{integer, number of colums (left to right on worktable) of labware}

\item{nrow}{integer, number of rows (back to front on worktable) of labware}

\item{zMove}{integer 0 - 5 type of movement: 0 = positioning with global
z-travel, 1 = positioning with local z-travel, 2 = x-move, 3 = y-move,
4 = z-move, 5 = Positioning with advanced z-travel (undocumented & untested..)}

\item{zTarget}{integer 0 - 4 z-position after move: 0 = z-travel
1 = z-dispense 2 = z-start 3 = z-max 4 = global z-travel}

\item{offset}{integer [-1000,1000] offset in mm added to z-position
(zTarget)}

\item{speed}{floating point [0.1,400] move speed in mm/s if zMove is 2,3 or 4}

\item{noOfLoopOptions}{integer 0 - 7 number of loop options required. If
non-zero, parameters \code{loopname, action, difference} are included
n times, where n = \code{noOfLoopOptions}}

\item{loopName}{loop name, \code{vector("character",noOfLoopOptions)}}

\item{action}{integer 0 - 3 loop action: 0 = vary the column,
1=vary the row, 2=vary the well 3=vary the rack,
\code{vector("character",noOfLoopOptions)}}

\item{difference}{integer number (“vary by”) which is used for the loop
action, \code{vector("character",noOfLoopOptions)}}

\item{arm}{integer 0 - 1, number of the MCA96 performing the action:
0 = MCA 1=MCA 2, should always be 0}
}
\description{
Move MCA to labware
}
\seealso{
Other advanced worklist commands: \code{\link{MCArelative}},
  \code{\link{MCAvector}}, \code{\link{adv_aspirate}},
  \code{\link{adv_dispense}},
  \code{\link{adv_gwl_comment}}, \code{\link{adv_mix}},
  \code{\link{adv_wash}}, \code{\link{fwCommand}},
  \code{\link{moveLiHa}}, \code{\link{startTimer}},
  \code{\link{waitTimer}}
}
