% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accFunctions.R
\name{parse_args}
\alias{parse_args}
\title{Parse Arguments for log}
\usage{
parse_args(arg_names, arg_types = NULL)
}
\arguments{
\item{arg_names}{names of arguments}

\item{arg_types}{One of NULL or a cols specification. See vignette("column-types", package = "readr") for more details.}
}
\description{
takes arguments to R script, parse them and pretty prints their values for log
}
\concept{accessory}
