% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot_F_histogram}
\alias{plot_F_histogram}
\title{Plot F-Distribution Method Results}
\usage{
plot_F_histogram(
  F_result,
  RD_obj,
  alpha = 0.01,
  binwidth = 0.1,
  log = FALSE,
  show_f_density = TRUE,
  ...
)
}
\arguments{
\item{F_result}{F_result object from thresh_F().}

\item{RD_obj}{Pre-computed RD_result object from \code{\link{compute_RD}}.}

\item{alpha}{Significance level used to compute RD threshold (default = 0.01 for 99th percentile).}

\item{binwidth}{Histogram bin width (default = 0.1).}

\item{log}{Logical. Apply log10 transformation to x-axis (default = TRUE).}

\item{show_f_density}{Logical. Show F-distribution curve overlay (default = TRUE).}

\item{...}{Additional arguments to to method-specific functions.}
}
\value{
A ggplot object.
}
\description{
Creates histogram of robust distances with F-distribution overlay and threshold.
}
