% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickoloFunctions.R
\name{plotColonies}
\alias{plotColonies}
\title{Plot Colonies from Pickolo}
\usage{
plotColonies(pickoloData, img_path, plotLabels = FALSE,
  colChange = NULL, line_size = 0.1, title_size = 4,
  label_size = 0.8, axis_text_size = 5, pars = pickoloData$pars)
}
\arguments{
\item{pickoloData}{List, output of \code{\link{readColonies}}}

\item{img_path}{path to the image of the plate}
}
\description{
Plot identified colonies over image from pickolo
}
\section{Output}{

   a \code{ggplot} plot
}

\seealso{
Other pickolo methods: \code{\link{readColonies}}
}
\concept{pickolo methods}
