% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_args.R
\name{quantile}
\alias{quantile}
\title{Quantile parameter documentation}
\arguments{
\item{quantile}{Numeric in (0,1) specifying the upper quantile for thresholding; the expected False Positive Rate for the chosen threshold.}
}
\description{
Quantile parameter documentation
}
