% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accFunctions.R
\name{rc_to_well}
\alias{rc_to_well}
\title{rc_to_well}
\usage{
rc_to_well(row, col, n_row, n_col)
}
\arguments{
\item{row}{a vector of integers or characters}

\item{col}{a vector of integers}

\item{n_row}{number of rows in labware}

\item{n_col}{number of columns in labware}
}
\description{
well number from row and column
}
\seealso{
Other general: \code{\link{generateMasks}},
  \code{\link{read_storex}}, \code{\link{shortestPath}}
}
\concept{general}
