% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_args.R
\name{RD_org_obj}
\alias{RD_org_obj}
\title{RD_org_obj parameter documentation}
\arguments{
\item{RD_org_obj}{Output list from \code{\link{compute_RD}} on the original data. Must contain $RD, $S_star, and $ind_incld.}
}
\description{
RD_org_obj parameter documentation
}
