% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwlCheckFunctions.R
\name{read_gwl}
\alias{read_gwl}
\title{Read in gwl file}
\usage{
read_gwl(gwl_path = file.choose(), .progress = TRUE)
}
\arguments{
\item{gwl_path}{path to gwl file, defaults to \code{file.choose()} to open dialog}

\item{.progress}{should a progress bar be shown?}
}
\description{
read in gwl file
}
\section{Output}{

   list of named character vectors
}

\seealso{
Other gwl_check: \code{\link{gwl_df_part}},
  \code{\link{read_gwl_df}}
}
