% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storeX_functions.R
\name{read_storex}
\alias{read_storex}
\title{Read Storex File}
\usage{
read_storex(storex_file)
}
\arguments{
\item{storex_file}{path to storex pos list}
}
\description{
Reads storex pos list to data.frame
}
\seealso{
Other general: 
\code{\link{generateMasks}()},
\code{\link{rc_to_well}()},
\code{\link{shortestPath}()}
}
\concept{general}
