% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickoloFunctions.R
\name{readColonies}
\alias{readColonies}
\title{Analyze colonies from Pickolo .csv}
\usage{
readColonies(csv_path, n_row = 16, row_start_coord = 185, row_d = 56.5,
  n_col = 24, col_start_coord = 155, col_d = 56.5)
}
\arguments{
\item{csv_path}{path to csv file from Pickolo}

\item{n_row}{number of rows spotted}

\item{row_start_coord}{at what coordinate on the picture do the rows start}

\item{row_d}{distance of rows}

\item{n_col}{number of cols spotted}

\item{col_start_coord}{at what coordinate on the picture do the cols start}

\item{col_d}{distance of cols}
}
\description{
Function to map colony growth to wells on 96well/384well plate
Defaults should work for 384 Well plate
}
\section{Output}{

   \code{list} with \code{data.frames} data and pars
}

\seealso{
Other pickolo methods: \code{\link{plotColonies}}
}
\concept{pickolo methods}
