% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iControlFunctions.R
\name{readInfiniteXML}
\alias{readInfiniteXML}
\title{read XML from Tecan Infinity Reader}
\usage{
readInfiniteXML(file)
}
\arguments{
\item{file}{path to XML file}
}
\description{
reads iControl XML files and returns a tibble
}
\section{Output}{

   \code{list} with data in longform format and parameters
}

