% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iControlFunctions.R
\name{readInfiniteXML2}
\alias{readInfiniteXML2}
\title{read XML from Tecan Infinity Reader}
\usage{
readInfiniteXML2(file)
}
\arguments{
\item{file}{path to XML file}
}
\description{
reads iControl XML file and returns a tibble
}
\section{Output}{

   \code{list} with data in longform format and parameters
}

