% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accFunctions.R
\name{shortestPath}
\alias{shortestPath}
\title{Shortest path for pin deactivation}
\usage{
shortestPath(pins, method = "two_opt")
}
\arguments{
\item{pins}{vector of the pins to be ordered}

\item{method}{method to use, see \code{\link{TSP::solve_TSP}}}
}
\description{
using the TSP package
}
\seealso{
Other general: 
\code{\link{generateMasks}()},
\code{\link{rc_to_well}()},
\code{\link{read_storex}()}
}
\concept{general}
