% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advGwlFunctions.R
\name{startTimer}
\alias{startTimer}
\title{Start Timer}
\usage{
startTimer(timer)
}
\arguments{
\item{timer}{Number of Timer to (re-)start}
}
\description{
Start Timer
}
\seealso{
Other advanced worklist commands: \code{\link{MCArelative}},
  \code{\link{MCAvector}}, \code{\link{adv_aspirate}},
  \code{\link{adv_dispense}},
  \code{\link{adv_gwl_comment}}, \code{\link{adv_mix}},
  \code{\link{adv_wash}}, \code{\link{fwCommand}},
  \code{\link{moveLiHa}}, \code{\link{moveMCA}},
  \code{\link{waitTimer}}
}
