% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procedures.R
\name{sterile_wash_c}
\alias{sterile_wash_c}
\title{Custom Sterile Wash procedure fro chemicals only}
\usage{
sterile_wash_c(
  wash_solution1 = "trough_EtOH",
  wash_solution1_liquidClass = "sterileWash_N_EtOH",
  wash_solution1_volume = 300,
  wash_solution2 = "trough_EtOH",
  wash_solution2_liquidClass = "sterileWash_N_EtOH",
  wash_solution2_volume = 400,
  wash_solution3 = "trough_H2O",
  wash_solution3_liquidClass = "sterileWash_N_H2O",
  wash_solution3_volume = 450
)
}
\arguments{
\item{wash_solution1}{RackLabel containing first wash solution}

\item{wash_solution1_liquidClass}{LiquidClass used to pipet first wash
solution}

\item{wash_solution1_volume}{volume used in first wash}

\item{wash_solution2}{RackLabel containing second wash solution}

\item{wash_solution2_liquidClass}{LiquidClass used to pipet
second wash solution}

\item{wash_solution2_volume}{volume used in second wash}

\item{wash_solution3}{RackLabel containing second wash solution}

\item{wash_solution3_liquidClass}{LiquidClass used to pipet
second wash solution}

\item{wash_solution3_volume}{volume used in third wash}
}
\description{
Washes & sterilizes all tips, hopefully. skips bleach
}
\seealso{
Other advanced worklist procedures: 
\code{\link{evowait}()},
\code{\link{iPTdeactivate}()},
\code{\link{iPTdip}()},
\code{\link{sterile_wash_TT}()},
\code{\link{sterile_wash_phage}()},
\code{\link{sterile_wash}()},
\code{\link{tiptouch}()},
\code{\link{washPinTool}()}
}
\concept{advanced worklist procedures}
