% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procedures.R
\name{sterile_wash_TT}
\alias{sterile_wash_TT}
\title{Short Sterile Wash procedure for Tip Touch Transfers}
\usage{
sterile_wash_TT(wash_solution1 = "trough_bleach",
  wash_solution1_liquidClass = "sterileWash_TT_bleach",
  wash_solution2 = "trough_EtOH",
  wash_solution2_liquidClass = "sterileWash_TT_EtOH",
  wash_solution3 = "trough_H2O",
  wash_solution3_liquidClass = "sterileWash_TT_H2O")
}
\arguments{
\item{wash_solution1}{RackLabel containing first wash solution}

\item{wash_solution1_liquidClass}{LiquidClass used to pipet first wash
solution}

\item{wash_solution2}{RackLabel containing second wash solution}

\item{wash_solution2_liquidClass}{LiquidClass used to pipet
second wash solution}

\item{wash_solution3}{RackLabel containing second wash solution}

\item{wash_solution3_liquidClass}{LiquidClass used to pipet
second wash solution}

\item{wash_rack}{Wash station used (default: wash1_shallow)}

\item{wash_waste}{Waste used (default: wash1_waste)}
}
\description{
Washes & sterilizes all tips, hopefully.
}
\seealso{
Other advanced worklist procedures: \code{\link{evowait}},
  \code{\link{iPTdeactivate}}, \code{\link{iPTdip}},
  \code{\link{sterile_wash_phage}},
  \code{\link{sterile_wash}}, \code{\link{tiptouch}},
  \code{\link{washPinTool}}
}
\concept{advanced worklist procedures}
