% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary.F_result}
\alias{summary.F_result}
\title{Summary method for Hardin & Rocke F results}
\usage{
\method{summary}{F_result}(object, ...)
}
\arguments{
\item{object}{An object of class "F_result" or "HR_result"}

\item{...}{Additional arguments (unused)}
}
\description{
Summary method for Hardin & Rocke F results
}
