% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary.MI_result}
\alias{summary.MI_result}
\title{Summary method for MI results}
\usage{
\method{summary}{MI_result}(object, ...)
}
\arguments{
\item{object}{An object of class "MI_result"}

\item{...}{Additional arguments (unused)}
}
\description{
Summary method for MI results
}
