% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary.SI_boot_result}
\alias{summary.SI_boot_result}
\title{Summary method for SI_boot results}
\usage{
\method{summary}{SI_boot_result}(object, ...)
}
\arguments{
\item{object}{An object of class "SI_boot_result"}

\item{...}{Additional arguments (unused)}
}
\description{
Summary method for SI_boot results
}
