% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_args.R
\name{symmetric}
\alias{symmetric}
\title{Symmetric parameter documentation}
\arguments{
\item{symmetric}{Logical. Use symmetric bounds for empirical rule.}
}
\description{
Symmetric parameter documentation
}
