% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_args.R
\name{thresh_result}
\alias{thresh_result}
\title{Thresh_result parameter documentation}
\arguments{
\item{thresh_result}{A threshold result object from any threshold method containing threshold information.}
}
\description{
Thresh_result parameter documentation
}
