% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_args.R
\name{threshold_method}
\alias{threshold_method}
\title{Threshold_method parameter documentation}
\arguments{
\item{threshold_method}{Character string; one of "all","SI","SI_boot","MI","MI_boot","F", "SHASH".}
}
\description{
Threshold_method parameter documentation
}
