% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_RD.R
\name{threshold_RD_old}
\alias{threshold_RD_old}
\title{Apply Outlier Detection Thresholding Methods}
\usage{
threshold_RD_old(
  name = NULL,
  data_matrix,
  alpha = 0.01,
  B = 1000,
  M = 100,
  k = 200,
  trans = "SHASH",
  cutoff = 3
)
}
\arguments{
\item{name}{Optional label for logging.}

\item{data_matrix}{T × V fMRI time series matrix.}

\item{alpha}{Quantile level for thresholding (default = 0.01).}

\item{B}{Number of bootstrap samples (default = 1000).}

\item{M}{Number of multiply imputed datasets (default = 50).}

\item{k}{Number of perturbation cycles per imputation (default = 100).}

\item{trans}{Character; transformation method for outlier detection (default = "SHASH").}

\item{cutoff}{Numeric; MAD-based outlier cutoff (default = 3).}
}
\value{
A list containing threshold values and intermediate diagnostics.
}
\description{
Applies five robust thresholding methods (HR, SI, SI_boot, MI, MI_boot)
after ICA-based univariate outlier detection and imputation.
}
