% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_args.R
\name{use_huber}
\alias{use_huber}
\title{Use_huber parameter documentation}
\arguments{
\item{use_huber}{Logical. If TRUE, apply Huber-like soft rejection.}
}
\description{
Use_huber parameter documentation
}
