% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advGwlFunctions.R
\name{waitTimer}
\alias{waitTimer}
\title{Wait for Timer}
\usage{
waitTimer(timer, timeSpan)
}
\arguments{
\item{timer}{Number of Timer to (re-)start}

\item{timeSpan}{duration [0.02 - 86400]}
}
\description{
Wait for Timer
}
\seealso{
Other advanced worklist commands: \code{\link{MCArelative}},
  \code{\link{MCAvector}}, \code{\link{adv_aspirate}},
  \code{\link{adv_dispense}},
  \code{\link{adv_gwl_comment}}, \code{\link{adv_mix}},
  \code{\link{adv_wash}}, \code{\link{fwCommand}},
  \code{\link{moveLiHa}}, \code{\link{moveMCA}},
  \code{\link{startTimer}}
}
