% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwlFunctions.R
\name{wash}
\alias{wash}
\title{Tecan Wash Command}
\usage{
wash(W = gwl$defaults$Wash)
}
\arguments{
\item{W}{"W" (or "W1"), "W2","W3","W4" or "WD" to select which wash scheme to use.}
}
\description{
The Wash Tips record washes the tip which was used by the preceeding Aspirate record. There are no parameters (the wash position, wash parameters are specified in the Worklist command).
}
\seealso{
Other basic worklist commands: \code{\link{aspirate}},
  \code{\link{dispense}}, \code{\link{distribute}},
  \code{\link{gwl_break}}, \code{\link{gwl_comment}},
  \code{\link{gwl_flush}}
}
\concept{basic worklist commands}
