% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accFunctions.R
\name{wellsFromANSII}
\alias{wellsFromANSII}
\title{Wells from ANSII string}
\usage{
wellsFromANSII(wellSelection)
}
\arguments{
\item{wellSelection}{well selection string}
}
\description{
calculates the selected wells from the ANSII string (The position starts with
  1 and increases from rear to front and left to right.)
}
\seealso{
Other accessory functions: \code{\link{ANSIIFromWells}},
  \code{\link{ANSIIfromVector}}
}
