% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{write.gwl}
\alias{write.gwl}
\title{Write worklist to file}
\usage{
write.gwl(gwl, print = FALSE, lineNumbers = TRUE, saveWT = FALSE,
  quietly = FALSE)
}
\arguments{
\item{gwl}{object of class gwl}

\item{print}{logical, should the worklist be printed in R? Defaults to \code{FALSE}}

\item{lineNumbers}{logical, should linenumbers be prepended to existing comments? Defaults to \code{FALSE}}

\item{saveWT}{logical, should the worktable be saved as a \code{txt} file?}
}
\description{
writes worklist to file, filename defined in \code{init}.
}
\seealso{
Other general methods: \code{\link{addToWorklist}},
  \code{\link{addToWorktable}}, \code{\link{init}}
}
\concept{general methods}
