% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHASH.R
\name{emprule_rob}
\alias{emprule_rob}
\title{Robust Empirical Rule Outlier Detection}
\usage{
emprule_rob(x, thr = 4, tail = c("both", "upper", "lower"))
}
\arguments{
\item{x}{Numeric vector.}

\item{thr}{Final threshold to flag outliers.}

\item{tail}{character string specifying which tail(s) to check for outliers.
Must be one of \code{"upper"}, \code{"lower"}, or \code{"both"}.
\itemize{
\item \code{"upper"}: detect upper-tail outliers only.
\item \code{"lower"}: detect lower-tail outliers only.
\item \code{"both"}: detect two-sided outliers (default).
}}
}
\value{
Logical vector: TRUE = outlier, FALSE = inlier.
}
\description{
Detects outliers using the median ± threshold × MAD rule.
}
\keyword{internal}
