% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHASH.R
\name{emprule_rob}
\alias{emprule_rob}
\title{Robust empirical rule}
\usage{
emprule_rob(x, thr = 4, use_huber = FALSE, upper_only = FALSE)
}
\arguments{
\item{x}{The data}

\item{thr}{MAD threshold}

\item{use_huber}{Use the Huber estimates for center and scale? Default:
\code{FALSE}.}

\item{upper_only}{Only consider upper threshold? Default: \code{FALSE}.}
}
\value{
Logical vector indicating whether each element in \code{x} is an
outlier (\code{TRUE} if an outlier).
}
\description{
Robust empirical rule outlier detection applicable to approximately Normal data
}
\keyword{internal}
