% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_RD.R
\name{threshold_RD}
\alias{threshold_RD}
\title{Comprehensive Outlier Detection Using Robust Distance Thresholding}
\usage{
threshold_RD(
  x,
  w = NULL,
  method = c("SI_boot", "MI", "MI_boot", "SI", "F", "SHASH", "all"),
  RD_obj = NULL,
  impute_method = "mean",
  cutoff = 4,
  trans = "SHASH",
  M = 50,
  k = 100,
  alpha = 0.01,
  quantile = 0.01,
  verbose = FALSE,
  boot_quant = 0.95,
  B = 1000
)
}
\arguments{
\item{x}{A numeric matrix or data frame of dimensions T × p (observations × variables).}

\item{w}{A numeric matrix (n_time × L) of low-kurtosis ICA components used as predictors (optional).}

\item{method}{Character string; one of "all","SI","SI_boot","MI","MI_boot","F", "SHASH".}

\item{RD_obj}{Pre-computed RD_result object from \code{\link{compute_RD}}.}

\item{impute_method}{Character string; imputation method for univariate outliers.}

\item{cutoff}{A numeric value indicating how many MADs away from the median to flag as outliers. The default value is set to be 4.}

\item{trans}{Character string; transformation method, one of "SHASH" or "robZ".}

\item{M}{Integer; number of multiply imputed datasets (default = 5).}

\item{k}{Integer; number of perturbation cycles per imputation (default = 10).}

\item{alpha}{Significance level used to compute RD threshold (default = 0.01 for 99th percentile).}

\item{quantile}{Numeric in (0,1) specifying the upper quantile for thresholding; the expected False Positive Rate for the chosen threshold.}

\item{verbose}{Logical; if TRUE, print progress messages.}

\item{boot_quant}{Numeric; confidence level for bootstrap confidence intervals (default = 0.95, for 95\% CI).}

\item{B}{Integer; number of bootstrap samples per imputed dataset (default = 1000).}
}
\value{
A list with:
\describe{
\item{thresholds}{Result from the specific threshold method, or list of all methods if "all".}
\item{RD_obj}{The robust distance object from compute_RD().}
\item{call}{The matched function call.}
}
}
\description{
Performs univariate outlier detection + imputation, robust distance, and multiple thresholding methods.
}
