% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_args.R
\name{w}
\alias{w}
\title{W parameter documentation}
\arguments{
\item{w}{A numeric matrix (n_time × L) of low-kurtosis ICA components used as predictors (optional).}
}
\description{
W parameter documentation
}
