% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procedures.R
\name{washPinTool}
\alias{washPinTool}
\title{Wash pin tool}
\usage{
washPinTool(t_wait = 2, t_dry = 90, n_dips = 3, reservoirs = c("Bleach",
  "ddH2O", "Isopropanol"), blots = c("Blot1", "BlotBleach", "BlotH2O"))
}
\arguments{
\item{t_wait}{integer, time (s) to wait on blotting paper}

\item{t_dry}{integer, time (s) over dryer}

\item{n_dips}{integer, number of dips in reagents}

\item{reservoirs}{vector of RackLabel of reservoirs to dip}

\item{blots}{vector of Racklaberl of blotting spots}
}
\description{
Procedure to wash pintool
}
\seealso{
Other advanced worklist procedures: \code{\link{evowait}},
  \code{\link{iPTdeactivate}}, \code{\link{iPTdip}},
  \code{\link{sterile_wash_TT}},
  \code{\link{sterile_wash_phage}},
  \code{\link{sterile_wash}}, \code{\link{tiptouch}}
}
\concept{advanced worklist procedures}
